requirements.txt
PyYAML>=6.0
networkx
matplotlib>=3.7